Make compute manager use nova-conductor for instance_update
Several tests require changes to ensure that the conductor service
is started (or set to local mode) before they ask the compute
manager to do instance updates.

Changes were required to some of the tests that check the value
of the timestamps by converting to strings. After passing the instances
through jsonutils.to_primitive(), the strings are a bit different.

Change-Id: I660e471178f3e5b36c4be27d47047fb4e4e1f851